requirements.txt
streamlit>=1.28.0
yagmail==0.15.293
python-dotenv==1.0.0
keyring==24.3.0
PyMuPDF==1.23.8
pandas==2.1.4
openpyxl==3.1.2
Pillow>=10.0.0
